Moving files to new destinations
 - manipulation/planner to multibody/inverse_kinematics
    - constraint_relaxing_ik
 - manipulation/planner to manipulation/util
    - robot_plan_interpolator

Moving includes updating all references (and namespaces as appropriate).
