+++ examples/kuka_iiwa_arm/lcm_plan_interpolator.cc
@@ -8,8 +8,8 @@ namespace drake {
 namespace examples {
 namespace kuka_iiwa_arm {
 
-using manipulation::planner::InterpolatorType;
-using manipulation::planner::RobotPlanInterpolator;
+using manipulation::util::InterpolatorType;
+using manipulation::util::RobotPlanInterpolator;
 using systems::DiagramBuilder;
 
 LcmPlanInterpolator::LcmPlanInterpolator(const std::string& model_path,
