+++ multibody/inverse_kinematics/constraint_relaxing_ik.h
@@ -10,8 +10,7 @@
 #include "drake/multibody/plant/multibody_plant.h"
 
 namespace drake {
-namespace manipulation {
-namespace planner {
+namespace multibody {
 
 /**
  * A wrapper class around the IK planner. This class improves IK's usability by
@@ -85,10 +84,9 @@ class ConstraintRelaxingIk {
                VectorX<double>* q_res);
 
   RandomGenerator rand_generator_;
-  multibody::MultibodyPlant<double> plant_;
-  multibody::BodyIndex end_effector_body_idx_{};
+  MultibodyPlant<double> plant_;
+  BodyIndex end_effector_body_idx_{};
 };
 
-}  // namespace planner
-}  // namespace manipulation
+}  // namespace multibody
 }  // namespace drake
